% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMEGA.R
\name{BIMEGA}
\alias{BIMEGA}
\title{BIMEGA: Bivariate Gaussian Mixture model for DNA methylation and gene expression data in cancer.}
\usage{
BIMEGA(METcancer, METnormal, MAcancer, MAnormal = NULL, listOfGenes = NULL,
  filter = TRUE, NoNormalMode = FALSE, OutputRoot = "")
}
\arguments{
\item{METcancer}{Matrix with the methylation data of cancer tissue with genes in rows and samples in columns.}

\item{METnormal}{Matrix with the normal methylation data of the same genes as in METcancer. Again genes in rows and samples in columns. The samples do not have to match with the cancer data.}

\item{MAcancer}{Gene expression data for cancer tissue with genes in rows and samples in columns.}

\item{MAnormal}{Gene expression data for normal tissue with genes in rows and samples in columns (optional data set).}

\item{listOfGenes}{Vector with genes names to be evaluated, names must coincide with the names of the rows of METcancer.}

\item{filter}{Logical indicating if the polynomial regression to select genes with significative relation between methylation and gene expression should be performed (default: TRUE).}

\item{NoNormalMode}{Logical indicating if the methylation states found in the cancer samples should be compared to the normal samples (default: TRUE).}

\item{OutputRoot}{Path to store the BIMEGA results object.}
}
\value{
A list with the following components:
\item{MethylationStates}{Matrix with for all genes the Methylation states 
    using DM-value (i.e. Differential methylation values) that are defined as 
    the methylation value with respect to the average normal methylation for a gene.}
\item{NrComponents}{The number of methylation states for each gene.}
\item{Models}{Bivariate Gaussian mixture model parameters for each gene.}
\item{MethylationDrivers}{Genes identified as transcriptionally predictive and differentially methylated by BIMEGA.}
\item{MixtureStates}{A list with the DM-values for each gene that is functional and differential.}
\item{Classifications}{Matrix with integers indicating to which mixture component each sample was assigned to for each gene.}
\item{FunctionalGenesResults}{Matrix with information on the polynomial regression fit for each driver gene.}
}
\description{
BIMEGA identifies DNA methylation driven genes by jointly modeling DNA methylation and gene expression data in cancer vs. normal and looking 
for homogeneous subpopulations. Matched gene expression data (e.g. from microarray technology or RNA sequencing) 
is also used to identify functional DNA methylation events by requiring a significant association between methylation 
and gene expression of a particular gene.
}
\examples{
# load the data sets needed for BIMEGA
data(METcancer)
data(METnormal)
data(MAcancer)
data(MAnormal)

# run BIMEGA on a small set of example data
BIMEGAresults <- BIMEGA(METcancer, METnormal, MAcancer, MAnormal)

}

