% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMEGA.R
\name{BIMEGA_Plot}
\alias{BIMEGA_Plot}
\title{The BIMEGA_Plot function.}
\usage{
BIMEGA_Plot(GeneName, MixtureModelResults, METcancer, MAcancer,
  METnormal = NULL, MAnormal = NULL, title = NULL)
}
\arguments{
\item{GeneName}{Name of the gene for which to create a BIMEGA plot.}

\item{MixtureModelResults}{List returned by BIMEGA function.}

\item{METcancer}{Matrix with the methylation data of cancer tissue with genes in rows and samples in columns.}

\item{MAcancer}{Gene expression data for cancer tissue with genes in rows and samples in columns.}

\item{METnormal}{Matrix with the normal methylation data of the same genes as in METcancer (optional). Again genes in rows and samples in columns.}

\item{MAnormal}{Gene expression data for normal tissue with genes in rows and samples in columns (optional).}

\item{title}{A title for the plot.}
}
\value{
BIMEGA plot, a scatterplot between DNA methylation and gene expression, showing the different mixture components identified.
}
\description{
Produces plots to represent BIMEGA's output.
}
\examples{
# load the data sets needed for BIMEGA
data(METcancer)
data(METnormal)
data(MAcancer)
data(MAnormal)

# run BIMEGA on a small set of example data
BIMEGAresults <- BIMEGA(METcancer, METnormal, MAcancer, MAnormal)

# Produce plots of differentially methylated genes
for (gene in BIMEGAresults$MethylationDrivers) {
     g <- BIMEGA_Plot(gene, BIMEGAresults, METcancer, MAcancer, METnormal, MAnormal)
     plot(g)
}

}

