% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMEGA.R
\name{fitRobustRegression_robustbase}
\alias{fitRobustRegression_robustbase}
\title{The fitRobustRegression_robustbase function}
\usage{
fitRobustRegression_robustbase(fitFormula)
}
\arguments{
\item{fitFormula}{formula for the model to be fitted.}
}
\value{
fitted robust regression model object.
}
\description{
Internal. Helper function to fit the robust regression.
}
\keyword{internal}

