% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMEGA.R
\name{getMvalues}
\alias{getMvalues}
\title{The getMvalues function}
\usage{
getMvalues(matr)
}
\arguments{
\item{matr}{matrix of beta values.}
}
\value{
matrix of M values
}
\description{
Internal. Converts Beta values into M values
}
\keyword{internal}

