% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{MAnormal}
\alias{MAnormal}
\title{Normal Gene expression data of glioblastoma patients}
\format{A numeric matrix with 13 rows (genes) and 10 columns (samples).}
\usage{
data(MAcancer)
}
\description{
Normal gene expression data of glioblastoma patients from the TCGA project and from
Noushmehr et al., 2010, and Sturm et al., 2012. A set of 13 genes that 
have been shown in the literature to be involved in 
differential methylation in glioblastoma were selected as an example to try out BIMEGA.
}
\references{
Cancer Genome Atlas Research Network. Comprehensive genomic characterization
defines human glioblastoma genes and core pathways. Nature. 2008 Oct 23;
455(7216):1061-8. doi: 10.1038/nature07385. Epub 2008 Sep 4. Erratum in: 
Nature. 2013 Feb 28;494(7438):506. PubMed PMID: 18772890; PubMed Central PMCID: PMC2671642.

Noushmehr,H. et al. (2010) Identification of a CpG island methylator phenotype that
defines a distinct subgroup of glioma. Cancer cell, 17, 510-22.

Sturm,D. et al. (2012) Hotspot Mutations in H3F3A and IDH1 Define Distinct 
Epigenetic and Biological Subgroups of Glioblastoma. Cancer Cell, 22, 425-437.
}
\seealso{
TCGA: The Cancer Genome Atlas: \url{http://cancergenome.nih.gov/}
}
\keyword{datasets}

