% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMEGA.R
\name{modelGeneExpression}
\alias{modelGeneExpression}
\title{The modelGeneExpression function}
\usage{
modelGeneExpression(METcancer, MAcancer, CovariateData = NULL)
}
\arguments{
\item{METcancer}{matrix with methylation data for cancer samples (genes in rows, samples in columns).}

\item{MAcancer}{matrix with gene expression data for cancer samples (genes in rows, samples in columns).}

\item{CovariateData}{vector (numeric or character) indicating a covariate to be included in the model to adjust for it. Not used in an standard run of BIMEGA.
It can be used if samples can from different tissue type, for example.}
}
\value{
matrix with R-square, degree of the best polynomial model and estimated coefficients only for the significant genes.
}
\description{
Model gene expression as a function of gene expression with a polynomial robust regression model (robust base package). 
Up to grade 3 models are evaluated and compared. R-squared and estimated coefficients are returned for the best model only for the genes with significant models.
}
\examples{
# load data sets
data(METcancer)
data(MAcancer)

# model gene expression
results <- modelGeneExpression(METcancer, MAcancer)

}

